Fixed remaining feedback
Almost only scss fixes. One minor php fix. Commit says changes in 500 files. This is because of line ending inconsistency since it was mostly developed on mac before, but this was done in windows. There is no project wise line ending rules and my attempt at normalizing all of them locally was of nu success. This however should change nothing